Fix deprecation warnings
dry-types 0.12.0 deprecated Types::Array#member and introduced a
replacement Types::Array#of in the same release [1].  Sigh.

This means that for dry-types <0.12, you must use Types::Array#member
because #of doesn't exist, but for dry-types >=0.12, you must use
Types::Array#of because #member emits a noisy deprecation warning.

This commit is a quick fix: it changes the gemspec to require >= 0.12,
and uses #of.

[1]: https://github.com/dry-rb/dry-types/blob/master/CHANGELOG.md#v0120-2017-09-15